Start implementing convert_to_student.py.
convert_to_student.py is a reimplementation of the conversion script

  //go/cmd/assign/assign --command=student

in pure Python. The initial implementation is developed with
pure Colab use case in mind (both authoring and student flow
happen in Colab, no extra backend or tooling is used),
so it may not be initially fully compatible with the existing script.